% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hartford_data}
\alias{hartford_data}
\title{Crimes and Spatial Risk Factors for Hartford, CT}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{https://openhartford-hartfordgis.opendata.arcgis.com/datasets/b226933a414046c498764d3b6821826f_2/explore}
}
\usage{
hartford_data
}
\description{
A dataset containing robberies in Hartford CT for the years
2018 - 2019. Spatial risk factors and city boundaries are included.
}
\keyword{datasets}
