% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_feature.R
\name{nearest_feature}
\alias{nearest_feature}
\title{Fast helper function for calculating nearest pairwise distances}
\usage{
nearest_feature(x, y)
}
\arguments{
\item{x}{X Value. Should be grid cell shapefile}

\item{y}{Y Value. Should be spatial predictors}
}
\description{
Fast helper function for calculating nearest pairwise distances
}
