% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eval.R
\name{plot_eval}
\alias{plot_eval}
\title{Plot one or more prediction evaluations}
\usage{
plot_eval(model_fit, test_data, eval = "pai", cutoff = 0.01)
}
\arguments{
\item{cutoff}{}
}
\description{
Plot one or more prediction evaluations
}
