% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_eval.R
\name{outcome_eval}
\alias{outcome_eval}
\title{Evaluate predictions from a gbm_fit model}
\usage{
outcome_eval(
  model_fit,
  test_data,
  eval = "pai",
  cutoff = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Evaluate predictions from a fitted gbm_fit model. Provides a variety of functions to estimate the out-of-sample
effectiveness of a prediction model, including the predictive accuracy index (PAI), the predictive efficiency index
(PEI), and the recapture rate index (RRI). At a minimum users must provide a fitted model object from the
\code{gbm_fit} function and a set of out-of-sample observations as 'test' data. These out-of-sample values should be
crimes or observations that were not used in the fitting the model.
}
