% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for analysis}
\usage{
prep_data(
  outcome,
  pred_var,
  region,
  gridsize,
  measure = "density",
  kernel = "auto"
)
}
\arguments{
\item{outcome}{Outcome variable as a point shapefile}

\item{pred_var}{Named list of shapefiles as spatial predictors.}

\item{region}{Polygon shapefile enclosing the study region}

\item{gridsize}{Size (in feet or meters) of the size of the spatial grid.}

\item{measure}{Types of measures for the predictor variables (either distance, density, or both). Defaults to distance.}

\item{kernel}{Named list of bandwidth distances if 'density' is chosen. Defaults to automatic selection.}
}
\value{
model list
}
\description{
This is a generic function that takes an outcome variable (i.e. a crime), a named list of predictor variables,
and a study region and converts it into a grid-based model object. Users can select whether to calculate nearest grid-cell distances,
densities, or grid cell counts. Function returns a (non-spatial) dataframe and a spatial (\code{sf}) polygon grid. The function is intended
to be used with the \code{lightgbm} wrapper 'lgm_prep' but can be used in any other statistical model (i.e. \code{ranger}, \code{glmnet}).
}
\examples{
data("hartford_data")

model_data <- 
prep_data(outcome = hartford_data[['robbery']],
          pred_var = hartford_data[c("bar","liquor","gas","pharmacy","retail")],
          region = hartford_data[['hartford']],
          gridsize = 200,
          measure = 'distance')
          
}
