% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for analysis}
\usage{
prep_data(
  outcome,
  pred_var,
  region,
  gridsize,
  measure = "density",
  kernel_bdw = "auto",
  count_vars = NULL
)
}
\arguments{
\item{outcome}{Outcome variable as a point shapefile}

\item{pred_var}{Named list of shapefiles as spatial predictors.}

\item{region}{Polygon shapefile enclosing the study region}

\item{gridsize}{Size (in feet or meters) of the size of the spatial grid.}

\item{measure}{Types of measures for the predictor variables (either distance, density, or both).
Defaults to distance.}

\item{kernel_bdw}{Either a numeric value or named list of bandwidth distances if 'density' is chosen. If a named list
is provided, the names should match the named predictor variables in the \code{pred_var} object. If nothing is provided,
defaults to automatic bandwidth selection via spatstat::bw.ppl. If speed is a concern, manually providing bandwidth
values is substantially faster than the automatic selection feature.}

\item{count_vars}{Vector of names of variables in \code{pred_var} to be counted. Default is all variables.}
}
\value{
model list
}
\description{
Prepares data for analysis. This is a generic function that takes an outcome variable (i.e. a crime), a named list
of predictor variables, and a study region and converts it into a grid-based model object.
Users can select whether to calculate nearest grid-cell distances,densities,counts, or all three. Function returns a
(non-spatial) dataframe and a spatial (\code{sf}) polygon grid. The function is intended to be used with the \code{xgboost}
wrapper \code{gbm_fit}, but can also used in any other statistical model (i.e. \code{ranger}, \code{glmnet}). A number of functions
are borrowed from code examples in Wheeler & Steenbeek, W. (2021).
}
\examples{
data("hartford_data")

# Prepping data for distance only
model_data <- prep_data(outcome = hartford_data$robbery,
                       pred_var = hartford_data[c('bar','nightclub','liquor','gas','pharmacy','restaurant')],
                       region = hartford_data$hartford,
                       gridsize = 200,
                       measure = 'distance')
                       
}
\references{
Wheeler, A. P., & Steenbeek, W. (2021). Mapping the risk terrain for crime using machine learning.
\emph{Journal of Quantitative Criminology}, 37(2), 445-480.
}
